+++ CHANGELOG.md
@@ -42,6 +42,7 @@ in 3.17. Use `PaymentSettings.defaultTransactionFlowStrategy` instead.
 - Fix giftcard code description - #13728 by @rafiwts
 - Change error message when denying a permission - #13334 by @rafiwts
 - Add filter by slugs to attribute choices - #13761 by @rafiwts
+- Add a new `product` field on `AssignedProductAttributeValue`. First part of a simplification of Attribute - Product relation from #12881. by @aniav
 
 # 3.15.0 [Unreleased]